Nilive-Bot/requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
flask==3.0.0
requests==2.31.0
Pillow==10.2.0
thefuzz==0.22.1
python-Levenshtein==0.25.0
